wip: move parse repo and pass info to github plugin
