+++ package.json
@@ -41,6 +41,7 @@
     },
     "devDependencies": {
         "@commitlint/cli": "^11.0.0",
+        "@monodeploy/git": "workspace:packages/git",
         "@monodeploy/test-utils": "link:./testUtils",
         "@monodeploy/types": "workspace:packages/types",
         "@tophat/commitizen-adapter": "^0.0.11",
