+++ packages/git/src/index.ts
@@ -1,2 +1,3 @@
 export * from './gitCommands'
 export * from './getCommitMessages'
+export * from './parseRepositoryProperty'
