+++ packages/plugin-github/package.json
@@ -25,8 +25,10 @@
     "@types/node": "^14.0.0"
   },
   "dependencies": {
+    "@monodeploy/git": "^0.1.8",
     "@monodeploy/logging": "^0.1.4",
     "@monodeploy/types": "^0.4.0",
-    "@octokit/core": "^3.4.0"
+    "@octokit/core": "^3.4.0",
+    "@octokit/plugin-throttling": "^3.4.1"
   }
 }
